Renames zapierother to zapier
