+++ packages/app-store/components.tsx
@@ -21,7 +21,7 @@ export const InstallAppButtonMap = {
   zoomvideo: dynamic(() => import("./zoomvideo/components/InstallAppButton")),
   office365video: dynamic(() => import("./office365video/components/InstallAppButton")),
   wipemycalother: dynamic(() => import("./wipemycalother/components/InstallAppButton")),
-  zapierother: dynamic(() => import("./zapierother/components/InstallAppButton")),
+  zapier: dynamic(() => import("./zapier/components/InstallAppButton")),
   jitsivideo: dynamic(() => import("./jitsivideo/components/InstallAppButton")),
   huddle01video: dynamic(() => import("./huddle01video/components/InstallAppButton")),
   metamask: dynamic(() => import("./metamask/components/InstallAppButton")),
