+++ packages/app-store/index.ts
@@ -15,7 +15,7 @@ import * as slackmessaging from "./slackmessaging";
 import * as stripepayment from "./stripepayment";
 import * as tandemvideo from "./tandemvideo";
 import * as wipemycalother from "./wipemycalother";
-import * as zapierother from "./zapierother";
+import * as zapier from "./zapier";
 import * as zoomvideo from "./zoomvideo";
 
 const appStore = {
@@ -37,7 +37,7 @@ const appStore = {
   wipemycalother,
   metamask,
   giphy,
-  zapierother,
+  zapier,
 };
 
 export default appStore;
