+++ packages/app-store/metadata.ts
@@ -14,7 +14,7 @@ import { metadata as slackmessaging } from "./slackmessaging/_metadata";
 import { metadata as stripepayment } from "./stripepayment/_metadata";
 import { metadata as tandemvideo } from "./tandemvideo/_metadata";
 import { metadata as wipemycalother } from "./wipemycalother/_metadata";
-import { metadata as zapierother } from "./zapierother/_metadata";
+import { metadata as zapier } from "./zapier/_metadata";
 import { metadata as zoomvideo } from "./zoomvideo/_metadata";
 
 export const appStoreMetadata = {
@@ -35,7 +35,7 @@ export const appStoreMetadata = {
   wipemycalother,
   metamask,
   giphy,
-  zapierother,
+  zapier,
 };
 
 export default appStoreMetadata;
