+++ packages/app-store/zapier/_metadata.ts
@@ -7,8 +7,8 @@ export const metadata = {
   description: _package.description,
   installed: true,
   category: "other",
-  imageSrc: "/api/app-store/zapierother/icon.svg",
-  logo: "/api/app-store/zapierother/icon.svg",
+  imageSrc: "/api/app-store/zapier/icon.svg",
+  logo: "/api/app-store/zapier/icon.svg",
   publisher: "Cal.com",
   rating: 0,
   reviews: 0,
