+++ packages/app-store/zapier/components/InstallAppButton.tsx
@@ -3,7 +3,8 @@ import type { InstallAppButtonProps } from "@calcom/app-store/types";
 import useAddAppMutation from "../../_utils/useAddAppMutation";
 
 export default function InstallAppButton(props: InstallAppButtonProps) {
-  const mutation = useAddAppMutation("zapier_other");
+  // @ts-ignore TODO: deprecate App types in favor of DB slugs
+  const mutation = useAddAppMutation("zapier");
 
   return (
     <>
