+++ packages/app-store/zapier/components/zapierSetup.tsx
@@ -4,13 +4,13 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { useState } from "react";
 
+import { useLocale } from "@calcom/lib/hooks/useLocale";
 import showToast from "@calcom/lib/notification";
 import { Button } from "@calcom/ui";
+import Loader from "@calcom/web/components/Loader";
+import { Tooltip } from "@calcom/web/components/Tooltip";
 
-import Loader from "../../../../apps/web/components/Loader";
-import { Tooltip } from "../../../../apps/web/components/Tooltip";
 import Icon from "./icon";
-import { useLocale } from "@calcom/lib/hooks/useLocale";
 
 interface IZapierSetupProps {
   trpc: any;
@@ -44,7 +44,7 @@ export default function ZapierSetup(props: IZapierSetupProps) {
 
   if (integrations.isLoading) {
     return (
-      <div className="absolute z-50 flex items-center w-full h-screen bg-gray-200">
+      <div className="absolute z-50 flex h-screen w-full items-center bg-gray-200">
         <Loader />
       </div>
     );
@@ -53,7 +53,7 @@ export default function ZapierSetup(props: IZapierSetupProps) {
   return (
     <div className="flex h-screen bg-gray-200">
       {showContent ? (
-        <div className="p-10 m-auto bg-white rounded">
+        <div className="m-auto rounded bg-white p-10">
           <div className="flex flex-row">
             <div className="mr-5">
               <Icon />
@@ -70,8 +70,8 @@ export default function ZapierSetup(props: IZapierSetupProps) {
               ) : (
                 <>
                   <div className="mt-1 text-xl">{t("your_unique_api_key")}</div>
-                  <div className="flex my-2 mt-3">
-                    <div className="w-full p-3 pr-5 mr-1 bg-gray-100 rounded">{newApiKey}</div>
+                  <div className="my-2 mt-3 flex">
+                    <div className="mr-1 w-full rounded bg-gray-100 p-3 pr-5">{newApiKey}</div>
                     <Tooltip content="copy to clipboard">
                       <Button
                         onClick={() => {
@@ -80,7 +80,7 @@ export default function ZapierSetup(props: IZapierSetupProps) {
                         }}
                         type="button"
                         className="px-4 text-base ">
-                        <ClipboardCopyIcon className="w-5 h-5 mr-2 text-neutral-100" />
+                        <ClipboardCopyIcon className="mr-2 h-5 w-5 text-neutral-100" />
                         {t("copy")}
                       </Button>
                     </Tooltip>
